Update my Guide to the Twin Cities document
I was recently informed that one of the links on here was out of date. When I went to update it, I noticed that this document was very out of date, and also contained some personal details I didn't mean to be included in it. This fixes links, descriptions, and adds a couple places I've found since the last update.